refactor: split tests into modular structure (16 files, 384 tests preserved)
